+++ packages/imperative/__tests__/src/packages/imperative/plugins/test_plugins/normal_plugin_3/lib/sample-plugin/cmd/foo/foo.handler.js
@@ -18,6 +18,7 @@ class FooHandler {
                 profile.size + " and duration = " + profile.duration;
             // impFileLogger.debug(successMsg);
             params.response.console.log(successMsg);
+            yield undefined;
         });
     }
 }