+++ packages/imperative/__tests__/src/packages/imperative/plugins/test_plugins/space in path plugin/lib/sample-plugin/cmd/foo/foo.handler.js
@@ -12,6 +12,7 @@ class FooHandler {
     process(params) {
         return __awaiter(this, void 0, void 0, function* () {
             params.response.console.log("Invoked sample-plugin foo handler");
+            yield undefined;
         });
     }
 }